isy-cli/pyproject.toml
[tool.poetry]
name = "isyflask-cli"
version = "0.6.0"
description = "Un cli para manejar proyectos de API con flask"
authors = ["David Cuy <david.cuy.sanchez@gmail.com>"]
readme = "README.md"
packages = [{include = "isyflask_cli"}]
homepage = "https://github.com/DavidCuy/easyflask-cli"
repository = "https://github.com/DavidCuy/easyflask-cli"
documentation = "https://github.com/DavidCuy/easyflask-cli"


[tool.poetry.scripts]
isyflask-cli = "isyflask_cli.cli:app"
isy = "isyflask_cli.cli:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.7.0"}
python-dotenv = "^0.20.0"
cookiecutter = "^2.1.1"
toml = "^0.10.2"
typing-extensions = "^4.9.0"
PyYAML = "~6.0.1"
setuptools = "~72.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
